move the validation logic from the download service to the validation service
