+++ src/xcode/ENA/ENA/Source/Services/HealthCertificateValidation/HealthCertificateValidationService.swift
@@ -6,6 +6,13 @@ import Foundation
 import OpenCombine
 import HealthCertificateToolkit
 
+import enum CertLogic.RuleType
+import enum CertLogic.CertificateType
+import class CertLogic.Rule
+import class CertLogic.ExternalParameter
+import class CertLogic.FilterParameter
+import class CertLogic.ValidationResult
+
 protocol HealthCertificateValidationProviding {
 	func validate(
 		healthCertificate: HealthCertificate,
@@ -37,7 +44,8 @@ final class HealthCertificateValidationService: HealthCertificateValidationProvi
 		signatureVerifier: SignatureVerification = SignatureVerifier(),
 		validationRulesAccess: ValidationRulesAccessing = ValidationRulesAccess(),
 		signatureVerifying: DCCSignatureVerifying,
-		dscListProvider: DSCListProviding
+		dscListProvider: DSCListProviding,
+		rulesDownloadService: RulesDownloadServiceProviding
 	) {
 		self.store = store
 		self.client = client
@@ -46,6 +54,7 @@ final class HealthCertificateValidationService: HealthCertificateValidationProvi
 		self.validationRulesAccess = validationRulesAccess
 		self.signatureVerifying = signatureVerifying
 		self.dscListProvider = dscListProvider
+		self.rulesDownloadService = rulesDownloadService
 	}
 		
 	// MARK: - Protocol HealthCertificateValidationProviding
@@ -101,7 +110,7 @@ final class HealthCertificateValidationService: HealthCertificateValidationProvi
 	
 	// MARK: - Private
 	
-	private var rulesDownloadService: ValidationRulesDownloadServiceProviding?
+	private var rulesDownloadService: RulesDownloadServiceProviding?
 	private let store: Store
 	private let client: Client
 	private let vaccinationValueSetsProvider: VaccinationValueSetsProviding
@@ -149,9 +158,7 @@ final class HealthCertificateValidationService: HealthCertificateValidationProvi
 					guard let self = self else { return }
 
 					Log.info("Successfully received value sets. Proceed with downloading acceptance rules...", log: .vaccination)
-					self.rulesDownloadService = RulesDownloadService(store: self.store, client: self.client)
-						
-					self.rulesDownloadService?.prepareAndDownloadValidationRules(
+					self.prepareAndDownloadValidationRules(
 						healthCertificate: healthCertificate,
 						arrivalCountry: arrivalCountry,
 						validationClock: validationClock,
@@ -161,6 +168,221 @@ final class HealthCertificateValidationService: HealthCertificateValidationProvi
 				}
 			)
 			.store(in: &subscriptions)
+	}
+	
+	// Validation Rules Preparation
+	
+	func prepareAndDownloadValidationRules(
+		healthCertificate: HealthCertificate,
+		arrivalCountry: Country,
+		validationClock: Date,
+		valueSets: SAP_Internal_Dgc_ValueSets,
+		completion: @escaping (Result<HealthCertificateValidationReport, HealthCertificateValidationError>) -> Void
+	) {
+		// 1. update/ download acceptance rules
+		self.rulesDownloadService?.downloadRules(
+			ruleType: .acceptance,
+			completion: { [weak self] result in
+				switch result {
+				case let .failure(error):
+					completion(.failure(error))
+				case let .success(acceptanceRules):
+					Log.info("Successfully downloaded/restored acceptance rules. Proceed with downloading invalidation rules...", log: .vaccination)
+					self?.downloadInvalidationRules(
+						healthCertificate: healthCertificate,
+						arrivalCountry: arrivalCountry,
+						validationClock: validationClock,
+						valueSets: valueSets,
+						acceptanceRules: acceptanceRules,
+						completion: completion
+					)
+				}
+			}
+		)
+	}
+	
+	private func downloadInvalidationRules(
+		healthCertificate: HealthCertificate,
+		arrivalCountry: Country,
+		validationClock: Date,
+		valueSets: SAP_Internal_Dgc_ValueSets,
+		acceptanceRules: [Rule],
+		completion: @escaping (Result<HealthCertificateValidationReport, HealthCertificateValidationError>) -> Void
+	) {
+		// 2. update/ download invalidation rules
+		self.rulesDownloadService?.downloadRules(
+			ruleType: .invalidation,
+			completion: { [weak self] result in
+				switch result {
+				case let .failure(error):
+					completion(.failure(error))
+				case let .success(invalidationRules):
+					Log.info("Successfully downloaded/restored invalidation rules. Proceed with assembling FilterParameter...", log: .vaccination)
+					self?.assembleCommonRules(
+						healthCertificate: healthCertificate,
+						arrivalCountry: arrivalCountry,
+						validationClock: validationClock,
+						valueSets: valueSets,
+						acceptanceRules: acceptanceRules,
+						invalidationRules: invalidationRules,
+						completion: completion
+					)
+				}
+			}
+		)
+	}
+	
+	private func assembleCommonRules(
+		healthCertificate: HealthCertificate,
+		arrivalCountry: Country,
+		validationClock: Date,
+		valueSets: SAP_Internal_Dgc_ValueSets,
+		acceptanceRules: [Rule],
+		invalidationRules: [Rule],
+		completion: @escaping (Result<HealthCertificateValidationReport, HealthCertificateValidationError>) -> Void
+	) {
+		// 3. Assemble common FilterParameter
+		
+		let mappedCertificateType = mapCertificateType(healthCertificate.type)
+		
+		// we must not set the region at the moment, so we set it to nil.
+		let filterParameter = FilterParameter(
+			validationClock: validationClock,
+			countryCode: arrivalCountry.id,
+			certificationType: mappedCertificateType,
+			region: nil
+		)
+		
+		Log.info("Successfully assembled common FilterParameter: \(private: filterParameter). Proceed with assembling common ExternalParameter...", log: .vaccination)
+		
+		// Assemble common ExternalParameter
+		
+		let mappedValueSets = mapValueSets(valueSet: valueSets)
+		let exp = healthCertificate.cborWebTokenHeader.expirationTime
+		let iat = healthCertificate.cborWebTokenHeader.issuedAt
+		
+		let externalParameter = ExternalParameter(
+			validationClock: validationClock,
+			valueSets: mappedValueSets,
+			exp: exp,
+			iat: iat,
+			issuerCountryCode: healthCertificate.cborWebTokenHeader.issuer,
+			kid: healthCertificate.keyIdentifier
+		)
+		
+		Log.info("Successfully assembled common ExternalParameter: \(private: externalParameter). Proceed with rule validation...", log: .vaccination)
+
+		validateRules(
+			healthCertificate: healthCertificate,
+			valueSets: valueSets,
+			acceptanceRules: acceptanceRules,
+			invalidationRules: invalidationRules,
+			filterParameter: filterParameter,
+			externalParameter: externalParameter,
+			completion: completion
+		)
+	}
+	
+	private func validateRules(
+		healthCertificate: HealthCertificate,
+		valueSets: SAP_Internal_Dgc_ValueSets,
+		acceptanceRules: [Rule],
+		invalidationRules: [Rule],
+		filterParameter: FilterParameter,
+		externalParameter: ExternalParameter,
+		completion: @escaping (Result<HealthCertificateValidationReport, HealthCertificateValidationError>) -> Void
+	) {
+		
+		// 4. apply acceptance and invalidation rules together and validate them
+		
+		let combinedRules = acceptanceRules + invalidationRules
+		
+		let validationRulesResult = validationRulesAccess.applyValidationRules(
+			combinedRules,
+			to: healthCertificate.digitalCovidCertificate,
+			filter: filterParameter,
+			externalRules: externalParameter
+		)
+
+		guard case let .success(validatedRules) = validationRulesResult else {
+			if case let .failure(error) = validationRulesResult {
+				Log.error("Could not validate rules.", log: .vaccination, error: error)
+				completion(.failure(.RULES_VALIDATION_ERROR(error)))
+			}
+			return
+		}
+		
+		Log.info("Successfully validated combined rules: \(private: validatedRules). Proceed with combined rule interpretation...", log: .vaccination)
+
+		interpretateRulesResult(
+			validatedRules: validatedRules,
+			completion: completion
+		)
+		
+	}
+	
+	private func interpretateRulesResult(
+		validatedRules: [ValidationResult],
+		completion: @escaping (Result<HealthCertificateValidationReport, HealthCertificateValidationError>) -> Void
+	) {
+		
+		// 5. interpret rules
+		
+		if validatedRules.allSatisfy({ $0.result == .passed }) {
+			// all rules has to be .passed
+			Log.info("Validation result is: validationPassed. Validation complete.", log: .vaccination)
+			completion(.success(.validationPassed(validatedRules)))
+		} else if validatedRules.contains(where: { $0.result == .open }) &&
+					!validatedRules.contains(where: { $0.result == .fail }) {
+			// At least one rule should contain now .open and there is no .fail
+			Log.info("Validation result is: validationOpen. Validation complete.", log: .vaccination)
+			completion(.success(.validationOpen(validatedRules)))
+		} else {
+			// At least one rule should contain now .fail
+			Log.info("Validation result is: validationFailed. Validation complete.", log: .vaccination)
+			completion(.success(.validationFailed(validatedRules)))
+		}
+	}
+
+	
+	// MARK: - ExternalParameters helpers
+	
+	// Internal for testing purposes
+	var allCountryCodes: [String] {
+		// This list of country codes comes from our techspec 1:1 as it is. Because we want to keep the copied format, we disable swiftlint here.
+		// swiftlint:disable line_length
+		// swiftlint:disable comma
+		return ["AD","AE","AF","AG","AI","AL","AM","AO","AQ","AR","AS","AT","AU","AW","AX","AZ","BA","BB","BD","BE","BF","BG","BH","BI","BJ","BL","BM","BN","BO","BQ","BR","BS","BT","BV","BW","BY","BZ","CA","CC","CD","CF","CG","CH","CI","CK","CL","CM","CN","CO","CR","CU","CV","CW","CX","CY","CZ","DE","DJ","DK","DM","DO","DZ","EC","EE","EG","EH","ER","ES","ET","FI","FJ","FK","FM","FO","FR","GA","GB","GD","GE","GF","GG","GH","GI","GL","GM","GN","GP","GQ","GR","GS","GT","GU","GW","GY","HK","HM","HN","HR","HT","HU","ID","IE","IL","IM","IN","IO","IQ","IR","IS","IT","JE","JM","JO","JP","KE","KG","KH","KI","KM","KN","KP","KR","KW","KY","KZ","LA","LB","LC","LI","LK","LR","LS","LT","LU","LV","LY","MA","MC","MD","ME","MF","MG","MH","MK","ML","MM","MN","MO","MP","MQ","MR","MS","MT","MU","MV","MW","MX","MY","MZ","NA","NC","NE","NF","NG","NI","NL","NO","NP","NR","NU","NZ","OM","PA","PE","PF","PG","PH","PK","PL","PM","PN","PR","PS","PT","PW","PY","QA","RE","RO","RS","RU","RW","SA","SB","SC","SD","SE","SG","SH","SI","SJ","SK","SL","SM","SN","SO","SR","SS","ST","SV","SX","SY","SZ","TC","TD","TF","TG","TH","TJ","TK","TL","TM","TN","TO","TR","TT","TV","TW","TZ","UA","UG","UM","US","UY","UZ","VA","VC","VE","VG","VI","VN","VU","WF","WS","YE","YT","ZA","ZM","ZW"]
+	}
+		
+
+	// Maps our valueSet on the value set of CertLogic. See https://github.com/corona-warn-app/cwa-app-tech-spec/blob/proposal/business-rules-dcc/docs/spec/dgc-validation-rules-client.md#data-structure-of-external-rule-parameters
+	// Internal for testing purposes
+	func mapValueSets(
+		valueSet: SAP_Internal_Dgc_ValueSets
+	) -> [String: [String]] {
+		var dictionary = [String: [String]]()
+		dictionary["country-2-codes"] = allCountryCodes
+		dictionary["covid-19-lab-result"] = valueSet.tcTr.items.map { $0.key }
+		dictionary["covid-19-lab-test-manufacturer-and-name"] = valueSet.tcMa.items.map { $0.key }
+		dictionary["covid-19-lab-test-type"] = valueSet.tcTt.items.map { $0.key }
+		dictionary["disease-agent-targeted"] = valueSet.tg.items.map { $0.key }
+		dictionary["sct-vaccines-covid-19"] = valueSet.vp.items.map { $0.key }
+		dictionary["vaccines-covid-19-auth-holders"] = valueSet.ma.items.map { $0.key }
+		dictionary["vaccines-covid-19-names"] = valueSet.mp.items.map { $0.key }
+		return dictionary
+	}
+
+	// Internal for testing purposes
+	func mapCertificateType(_ certificateType: HealthCertificate.CertificateType) -> CertLogic.CertificateType {
+		switch certificateType {
+		case .vaccination:
+			return .vaccination
+		case .test:
+			return .test
+		case .recovery:
+			return .recovery
+		}
 	}
 	
 	// Internal for testing purposes
